frontend/bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Telegram bot frontend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.10",
    "httpx>=0.27.0",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "fuzzywuzzy>=0.18.0",
    "pandas>=2.2.3",
    "python-levenshtein>=0.26.1",
]

[project.scripts]
bot = "bot:entrypoint"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
